backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
google-api-python-client==2.111.0
google-auth==2.26.1
pandas==2.1.4
gunicorn==21.2.0
